rearrange sound initialization a little
Pulseaudio intermittently refuses to connect a new sound stream,
and immediately restarting the program gets it to connect again,
with the exact same parameters.

Move audio initialization to its own function, retry a few times
when initialization fails, and print the exception we got from
the SoundCard library to help debugging the issue.

Signed-off-by: Rik van Riel <riel@surriel.com>