fix rigctl mode set errors
Setting the mode with bandwidth to the rigctl backend could throw
an error, since splitting the command string returned a string, not
a number like expected for the call to set_mod.

Also pass None for the low cut, since that makes set_mod do the right
thing when switching between USB, LSB, and other modes.

Signed-off-by: Rik van Riel <riel@surriel.com>