reconnect with the same frequency after getting disconnected
Make kiwiclient remember what frequency it is on with the same
variable name in both kiwiclientd.py and kiwi/client.py, and
have _setup_rx_params use the remembered frequency when reconnecting.

Tested by starting the script, changing the frequency through the
rigctld interface, kicking the connection from the kiwisdr admin
page, and watching it reconnect on the same frequency.

Signed-off-by: Rik van Riel <riel@surriel.com>